Remove ChangePasswordAdviceRepository
Given that this works well with only a service, it's worth considering
whether we need a repository at all.

One concern is whether admins want users to be immediately advised or
if they want them to be advised the next time that they log in.

In this commit, they are advised immediately since it is stored in a backend
by default instead of in the session, which an admin cannot update.