Split ChangePassword Service and Repository
This commit simplifies the relationship between the service and
the repository by recognizing that the only time they are both
used is when the password is being changed.

This makes the password processing filter more complicated.
But I like this for now because it means that by default, users
are advised at login time and not in the middle of their session,
which I think is an easier default to adopt. Changing it to be
notified immediately is as simple as publishing a repository
that is backed by the service instead of the session.